+++ core/src/main/java/org/apache/struts/chain/commands/AbstractSelectInput.java
@@ -22,6 +22,7 @@ package org.apache.struts.chain.commands;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.apache.struts.action.Action;
 import org.apache.struts.chain.contexts.ActionContext;
 import org.apache.struts.config.ActionConfig;
 import org.apache.struts.config.ForwardConfig;
@@ -66,16 +67,16 @@ public abstract class AbstractSelectInput extends ActionCommandBase {
         ModuleConfig moduleConfig = actionConfig.getModuleConfig();
 
         // Cache an ForwardConfig back to our input page
-        ForwardConfig forwardConfig;
+        ForwardConfig forwardConfig = null;
         String input = actionConfig.getInput();
 
         if (moduleConfig.getControllerConfig().getInputForward()) {
             if (LOG.isTraceEnabled()) {
                 LOG.trace("Finding ForwardConfig for '" + input + "'");
             }
-            forwardConfig = actionConfig.findForwardConfig(input);
+            forwardConfig = inputForward(actionConfig, moduleConfig, input);
             if (forwardConfig == null) {
-                forwardConfig = moduleConfig.findForwardConfig(input);
+                LOG.error(getErrorMessage(actionCtx, actionConfig));
             }
         } else {
             if (LOG.isTraceEnabled()) {
@@ -106,4 +107,48 @@ public abstract class AbstractSelectInput extends ActionCommandBase {
      */
     protected abstract ForwardConfig forward(ActionContext context,
         ModuleConfig moduleConfig, String uri);
+
+    /**
+     * <p> Retrieve error message from context. </p>
+     *
+     * @param context      The <code>Context</code> for the current request
+     * @param actionConfig The current action mapping
+     * @return error message
+     */
+    protected abstract String getErrorMessage(ActionContext context,
+        ActionConfig actionConfig);
+
+    /**
+     * Attempts to resolve the input as a {@link ForwardConfig} attribute.
+     * This method should only invoked if the Controller has its
+     * <code>inputForward</code> property set to <code>true</code>.
+     * If the input parameter is specified, use that, otherwise try
+     * to find one in the mapping or the module under the standard
+     * conventional <code>input</code> name.
+     *
+     * @param actionConfig the config for the target action
+     * @param moduleConfig the config for the module of the action
+     * @param input the name of the input
+     * @return ForwardConfig representing destination
+     * @see Action#INPUT
+     */
+    protected ForwardConfig inputForward(ActionConfig actionConfig,
+            ModuleConfig moduleConfig, String input) {
+        ForwardConfig forwardConfig;
+
+        if (input != null) {
+            forwardConfig = actionConfig.findForwardConfig(input);
+            if (forwardConfig == null) {
+                forwardConfig = moduleConfig.findForwardConfig(input);
+            }
+        } else {
+            forwardConfig = actionConfig.findForwardConfig(Action.INPUT);
+            if (forwardConfig == null) {
+                forwardConfig = moduleConfig.findForwardConfig(Action.INPUT);
+            }
+        }
+
+        return forwardConfig;
+    }
+
 }